+++ BE/src/auth/auth.controller.ts
@@ -3,14 +3,10 @@ import { AuthService } from './auth.service';
 import { Controller, Get, UseGuards, Req, Res } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { Request, Response } from 'express';
-import { UsersService } from 'src/users/users.service';
 
 @Controller('auth')
 export class AuthController {
-  constructor(
-    private readonly authService: AuthService,
-    private readonly usersService: UsersService,
-  ) {}
+  constructor(private readonly authService: AuthService) {}
 
   @Get('')
   @UseGuards(AuthGuard('google'))
@@ -18,18 +14,9 @@ export class AuthController {
 
   @Get('google')
   @UseGuards(AuthGuard('google'))
-  async googleAuthRedirect(@Req() req) {
+  googleAuthRedirect(@Req() req) {
     const user = req.user;
-    const prevUser = await this.usersService.findUserByEmail(user.email);
-    if (!prevUser) {
-      const userEntity = {
-        nickname: 'test',
-        google_id: user.email,
-        image_url: '',
-      };
-      await this.usersService.createUser(userEntity);
-    }
-    return { message: '로그인 성공' };
+    return this.authService.loginWithGoogle(user);
   }
 
   @Get('logout')
